(merge) Resolve merge conflicts
