--- src/app/pages/tab1/tab1.page.html
@@ -1,22 +1,0 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <ion-title>
-      Near incidents
-    </ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content [fullscreen]="true">
-  <ion-list>
-    <ion-item *ngFor=" let marker of markers">
-      <ion-avatar slot="start">
-        <img src="https://s2.ppllstatics.com/elcomercio/www/multimedia/202301/21/media/cortadas/80120341--1248x968.jpg" />
-    </ion-avatar>
-    <ion-label>
-        <h2>{{marker.title}}</h2>
-        <h3>{{marker.snippet}}</h3>
-        <p>Coordinates {{marker.coordinate.lat}} {{marker.coordinate.lng}}</p>
-    </ion-label>
-    </ion-item>
-</ion-list>
-</ion-content>
