--- src/app/pages/tab2/tab2.page.html
@@ -1,16 +1,0 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <ion-title>
-      Incidents map
-    </ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content [fullscreen]="true">
-  <ion-fab slot="fixed" vertical="top" horizontal="start">
-    <ion-fab-button  size="small" (click)="changeLayer()">
-      <ion-icon name="layers-outline"></ion-icon>
-    </ion-fab-button>
-  </ion-fab>
-  <capacitor-google-map id="map"></capacitor-google-map>
-</ion-content>
