--- src/app/pages/tab2/tab2.page.scss
@@ -1,18 +1,0 @@
-capacitor-google-map {
-    display: inline-block;
-    width: 100%;
-    height: 99.5%;
-  }
-
-  ion-content {
-    --background: none;
-  }
-
-  ion-fab-button{
-    --background: white;
-    --color: black;
-  }
-  
-  body {
-    background: transparent;
-  }
