--- src/app/pages/tab3/tab3.page.ts
@@ -1,16 +1,0 @@
-import { Component } from '@angular/core';
-import { ExploreContainerComponent } from '../../explore-container/explore-container.component';
-import { IonicModule } from '@ionic/angular';
-
-@Component({
-    selector: 'app-tab3',
-    templateUrl: 'tab3.page.html',
-    styleUrls: ['tab3.page.scss'],
-    standalone: true,
-    imports: [IonicModule, ExploreContainerComponent]
-})
-export class Tab3Page {
-
-  constructor() {}
-
-}
