--- src/app/pages/tabs/tabs.page.html
@@ -1,20 +1,0 @@
-<ion-tabs>
-
-  <ion-tab-bar slot="bottom">
-    <ion-tab-button tab="list" href="/tabs/list">
-      <ion-icon aria-hidden="true" name="menu-outline"></ion-icon>
-      <!-- <ion-label>Tab 1</ion-label> -->
-    </ion-tab-button>
-
-    <ion-tab-button tab="map" href="/tabs/map">
-      <ion-icon aria-hidden="true" name="location-outline"></ion-icon>
-      <!-- <ion-label>Tab 2</ion-label> -->
-    </ion-tab-button>
-
-    <ion-tab-button tab="profile" href="/tabs/profile">
-      <ion-icon aria-hidden="true" name="person-outline"></ion-icon>
-      <!-- <ion-label>Tab 3</ion-label> -->
-    </ion-tab-button>
-  </ion-tab-bar>
-
-</ion-tabs>
